requirements.txt
slackclient>=1.0.5
slackeventsapi>=0.1
google-api-python-client
